wip: transfer responsibility to talk to UI Control service to Flow Control Service
